+++ cli/src/tool/tool_register_offchain.rs
@@ -18,7 +18,7 @@ use {
     },
     nexus_sdk::{
         move_bindings::{
-            online_payment::gas::OverGas,
+            gas::gas::OverGas,
             primitives::owner_cap::CloneableOwnerCap,
             registry::tool_registry::OverTool,
             struct_tag_matches,