+++ sdk/src/transactions/tool.rs
@@ -1,7 +1,7 @@
 use {
     crate::{
         move_bindings::{
-            online_payment::gas as gas_binding,
+            gas::gas as gas_binding,
             registry::{
                 registered_key_verifier as registered_key_verifier_binding,
                 tool_registry as tool_registry_binding,
@@ -667,7 +667,7 @@ mod tests {
 
     fn nexus_objects() -> NexusObjects {
         NexusObjects {
-            online_payment_pkg_id: addr("0x13"),
+            gas_pkg_id: addr("0x13"),
             workflow_pkg_id: addr("0x1"),
             scheduler_pkg_id: addr("0x11"),
             primitives_pkg_id: addr("0x2"),
@@ -695,7 +695,7 @@ mod tests {
             primitives_original_pkg_id: None,
             interface_original_pkg_id: None,
             registry_original_pkg_id: None,
-            online_payment_original_pkg_id: None,
+            gas_original_pkg_id: None,
             workflow_original_pkg_id: None,
             scheduler_original_pkg_id: None,
         }